+++ slim_cli/__init__.py
@@ -1,1 +1,1 @@
-version = '0.0.X'
+version = '0.3.X'